Complete the Odin Recipe Project
This was my second attempt at this project as I
took some time off so I just wanted to make sure
I could get the html on the page and working as
needed. I did it rather quickly without realizing
I did not do any commits. I will work on the
commits on the next project.